lucet-idl: witx use resolution moved to toplevel.rs
